Moved OriginPlacement and SubLevels into own files
